实现了 string_extend 函数, 未测试 修改了 chain_table_clear 函数传入空管理器时出现异常的bug, 教训: 非固定的默认参数未必是好的

+++ inc/chain_table.h
@@ -53,6 +53,6 @@ int chain_table_insert(ChainTableManager *manager, size_t element_size, bool is_
 
 int string_read(ChainTableManager *string, char *dest, int max_length);
 
-int string_extend(ChainTableManager*string, const char * source, int source_length, uint16_t node_length);
+int string_extend(ChainTableManager*string, const char * source, int64_t source_length, uint16_t node_length);
 
 #endif //ASSESSMENT_PIZZA_CHAIN_TABLE_H
